Now closed VS. closed. Download images with callback
