+++ app/src/main/res/values-de/strings.xml
@@ -72,7 +72,8 @@
 
     <string name="opens_at">Öffnet um %1$s Uhr</string>
     <string name="closes_at">Schließt um %1$s Uhr</string>
-    <string name="closed">Jetzt geschlossen</string>
+    <string name="closed">geschlossen</string>
+    <string name="now_closed">Jetzt geschlossen</string>
 
     <string name="available_at">Verfügbar in %1$s</string>
     <string name="available_since">Noch %1$s</string>
