+++ app/src/main/res/values/strings.xml
@@ -65,7 +65,8 @@
 
     <string name="opens_at">Opens at %1$s</string>
     <string name="closes_at">Closes at %1$s</string>
-    <string name="closed">Now closed</string>
+    <string name="closed">closed</string>
+    <string name="now_closed">Now closed</string>
 
     <string name="available_at">Disponsable in %s$1</string>
     <string name="available_since">Disponsable for %s$1</string>
